Updated and restructured repository.
A freebsd branch was created, since my linux machine is set
up differently.  The master branch is likely to be replaced with
an archlinux branch in a future commit.

The files within this repository were moved into directories
corresponding to GNU Stow packages.  GNU Stow is convenient for
creating symlinks into a clone of this repository.